requirements.txt
streamlit
qdrant-client
sentence-transformers
PyMuPDF
python-docx
transformers
accelerate
